+++ js/calculators/mara.js
@@ -9,7 +9,7 @@
  */
 
 import { createCache } from '../utils/simulation.js';
-import { registerCalculator } from '../utils/calculatorBase.js';
+import { registerCalculator, updateCalculatorIfActive } from '../utils/calculatorBase.js';
 import { renderHeroStats, generateSampleRevealsHTML } from '../utils/components.js';
 import { shuffleDeck, renderCardBadge, createCollapsibleSection, TYPE_COLORS, buildDeckFromCardData } from '../utils/sampleSimulator.js';
 import * as OpponentState from '../utils/opponentState.js';
@@ -513,7 +513,7 @@ export function init() {
         for (const key of Object.keys(stableSamples)) {
             stableSamples[key] = [];
         }
-        updateUI();
+        updateCalculatorIfActive('mara', updateUI);
     });
 
     registerCalculator({
@@ -530,8 +530,6 @@ export function init() {
             if (sampleBtn) {
                 sampleBtn.addEventListener('click', refreshSamples);
             }
-
-            updateUI();
         }
     });
 }